update extension function with primitive type modification for ServiceAction.Play.startTime

+++ sphinx/service/features/media-player/feature-service-media-player-android/src/main/java/chat/sphinx/feature_service_media_player_android/util/Extensions.kt
@@ -45,8 +45,8 @@ internal inline fun Intent.toServiceActionPlay(): UserAction.ServiceAction.Play?
         }
     }
 
-    val startTime: Long = getLongExtra("START_TIME", -1L).let {
-        if (it == -1L) {
+    val startTime: Int = getIntExtra("START_TIME", -1).let {
+        if (it == -1) {
             return null
         } else {
             it
